Add event switch system & handle HTTP responses
The event switch abstracts over waiting for any of multiple events. It
will remove all listeners after any event is emitted.

We also consider any HTTP responses to be errors because we are trying
to connect as a WebSocket. If we can connect as a WebSocket, the upgrade
event will be emitted, not the response event. So if we see response we
should throw. This should handle the Discord gateway being down with a
Cloudflare error page or whatever. DiscordConnector should retry
connecting if such a situation occurs in the future.